LayoutTests/fast/forms/label/labels-remove-parent-label.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/label-test-util.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Test that the labels list of a form control is dynamically updated when removing a label as parent for that form control. ');

var tester = setupLabelsTest(withParentLabel);
for (var name in tester.dataSet) {
    var data = tester.getLabelableElementData(name);
    if (data == null)
        continue;

    debug('Check ' + data.name);
    var element = data.element;
    shouldBe('element.labels.length', '1');

    tester.outerElement.appendChild(element);
    shouldBe('element.labels.length', '0');
}
</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
